requirements.txt
asgiref==3.4.1
Django==3.2.8
django-restframework==0.0.1
djangorestframework==3.12.4
Pillow==8.4.0
Pygments==2.10.0
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2